+++ emp-service-model/src/test/java/test/org/fugerit/java/emp/sm/service/TestServiceMessage.java
@@ -4,14 +4,22 @@ import org.fugerit.java.emp.sm.service.ServiceMessage;
 import org.junit.Assert;
 import org.junit.Test;
 
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 public class TestServiceMessage {
 
-	private static final String TEST_CODE = "401001";
+	private static final String TEST_CODE = "500001";
 	private static final String TEST_MESSAGE = "Test error";
+	
+	public static ServiceMessage newDefaultMessage( String severity ) {
+		return new ServiceMessage(TEST_CODE, severity, TEST_MESSAGE);
+	}
 	
 	@Test
 	public void testMessageAllArgs() {
-		ServiceMessage message = new ServiceMessage( TEST_CODE , ServiceMessage.SEVERITY_ERROR, "Test error");
+		ServiceMessage message = new ServiceMessage( TEST_CODE , ServiceMessage.SEVERITY_ERROR, TEST_MESSAGE);
+		log.info( "message : {}", message );
 		Assert.assertEquals( TEST_CODE , message.getCode() );
 		Assert.assertEquals( TEST_MESSAGE , message.getText() );
 		Assert.assertEquals( ServiceMessage.SEVERITY_ERROR , message.getSeverity() );